requirements.txt
python-dotenv==1.1.1
sentence-transformers==5.0.0
einops==0.8.1

pymongo==4.13.2

langchain==0.3.27
langchain-community==0.3.27
pypdf==5.8.0

langchain-google-genai==2.1.8
